+++ src/types.rs
@@ -5,54 +5,60 @@ use std::net::TcpStream;
 use std::time::Duration;
 use tungstenite::{stream::MaybeTlsStream, WebSocket};
 
+#[derive(Debug)]
+/// The connection to the EventSub server with Twitch, which contains a socket, a session ID, and
+/// the vector of handled messages (to avoid handling duplicates).
+pub struct Session {
+    /// The socket which is connected to Twitch's EventSub WebSocket server.
+    pub socket: Socket,
+    /// The session ID Twitch returns with the `Welcome` message. Initially empty String.
+    pub id: String,
+    /// The `handled` vector contains the message IDs of those messages which have already been
+    /// handled, to avoid taking action twice when Twitch repeats their notification.
+    pub handled: Vec<String>,
+}
+
+/// This layered type is [`tungstenite`](https://crates.io/crates/tungstenite)'s WebSocket connection.
 pub type Socket = WebSocket<MaybeTlsStream<TcpStream>>;
 
-pub trait MessageFields {
-    fn id(&self) -> String;
+#[derive(Serialize, Deserialize, Debug)]
+#[serde(untagged)]
+pub enum TwitchMessage {
+    Notification(Notification),
+    Welcome(Welcome),
+    Reconnect(Reconnect),
+    Revocation(Revocation),
+    Keepalive(Keepalive),
 }
 
-#[derive(Debug)]
-pub struct Session {
-    socket: Socket,
-    id: String,
-    handled: Vec<String>,
+#[derive(Deserialize, Serialize, Debug)]
+pub struct Notification {
+    metadata: SubscriptionMetadata,
+    payload: NotificationPayload,
 }
 
-impl Session {
-    pub fn socket(&mut self) -> &mut Socket {
-        &mut self.socket
-    }
-
-    pub fn handled(&mut self) -> &mut Vec<String> {
-        &mut self.handled
-    }
-
-    pub fn set_id(&mut self, id: String) {
-        self.id = id;
-    }
-
-    pub fn set_keepalive(&mut self, keepalive: u64) -> Result<(), std::io::Error> {
-        let stream = match self.socket().get_mut() {
-            MaybeTlsStream::NativeTls(stream) => stream.get_mut(),
-            MaybeTlsStream::Plain(stream) => stream,
-            _ => unreachable!("Stream has to always be either TLS or plain"),
-        };
-        stream
-            // Allow a short grace period by adding one second to the reported keepalive
-            // timeout
-            .set_read_timeout(Some(Duration::from_secs(keepalive + 1)))?;
-        Ok(())
-    }
+#[derive(Deserialize, Serialize, Debug)]
+pub struct Welcome {
+    metadata: GenericMetadata,
+    payload: WelcomeSessionPayload,
 }
 
-impl Session {
-    pub fn new(socket: Socket) -> Session {
-        Session {
-            socket,
-            id: String::new(),
-            handled: vec![],
-        }
-    }
+#[derive(Deserialize, Serialize, Debug)]
+pub struct Reconnect {
+    metadata: GenericMetadata,
+    payload: ReconnectSessionPayload,
+}
+
+#[derive(Deserialize, Serialize, Debug)]
+pub struct Keepalive {
+    metadata: GenericMetadata,
+    payload: Value,
+}
+
+#[derive(Deserialize, Serialize, Debug)]
+pub struct Revocation {
+    metadata: SubscriptionMetadata,
+    payload: SubscriptionPayload,
 }
 
 #[derive(Deserialize, Serialize, Debug)]
@@ -91,40 +97,6 @@ pub struct ReconnectSessionPayload {
 }
 
 #[derive(Deserialize, Serialize, Debug)]
-pub struct Welcome {
-    metadata: GenericMetadata,
-    payload: WelcomeSessionPayload,
-}
-
-impl Welcome {
-    pub fn session_id(&self) -> &str {
-        &self.payload.session.id
-    }
-
-    pub fn keepalive(&self) -> &Number {
-        &self.payload.session.keepalive_timeout_seconds
-    }
-}
-
-impl MessageFields for Welcome {
-    fn id(&self) -> String {
-        self.metadata.message_id.clone()
-    }
-}
-
-#[derive(Deserialize, Serialize, Debug)]
-pub struct Keepalive {
-    metadata: GenericMetadata,
-    payload: Value,
-}
-
-impl MessageFields for Keepalive {
-    fn id(&self) -> String {
-        self.metadata.message_id.clone()
-    }
-}
-
-#[derive(Deserialize, Serialize, Debug)]
 pub struct SubscriptionMetadata {
     message_id: String,
     message_type: String,
@@ -151,15 +123,64 @@ pub struct NotificationPayload {
     event: Value,
 }
 
-#[derive(Deserialize, Serialize, Debug)]
-pub struct Notification {
-    metadata: SubscriptionMetadata,
-    payload: NotificationPayload,
+impl Session {
+    pub fn new(socket: Socket) -> Session {
+        Session {
+            socket,
+            id: String::new(),
+            handled: vec![],
+        }
+    }
+
+    pub fn set_keepalive(&mut self, keepalive: u64) -> Result<(), std::io::Error> {
+        let stream = match self.socket.get_mut() {
+            MaybeTlsStream::NativeTls(stream) => stream.get_mut(),
+            MaybeTlsStream::Plain(stream) => stream,
+            _ => unreachable!("Stream has to always be either TLS or plain"),
+        };
+        stream
+            // Allow a short grace period by adding one second to the reported keepalive
+            // timeout
+            .set_read_timeout(Some(Duration::from_secs(keepalive + 1)))?;
+        Ok(())
+    }
+}
+
+impl Welcome {
+    pub fn session_id(&self) -> &str {
+        &self.payload.session.id
+    }
+
+    pub fn keepalive(&self) -> &Number {
+        &self.payload.session.keepalive_timeout_seconds
+    }
 }
 
 impl Notification {
     pub fn payload(&self) -> &NotificationPayload {
         &self.payload
+    }
+}
+
+impl Reconnect {
+    pub fn reconnect_url(&self) -> &String {
+        &self.payload.session.reconnect_url
+    }
+}
+
+pub trait MessageFields {
+    fn id(&self) -> String;
+}
+
+impl MessageFields for Welcome {
+    fn id(&self) -> String {
+        self.metadata.message_id.clone()
+    }
+}
+
+impl MessageFields for Keepalive {
+    fn id(&self) -> String {
+        self.metadata.message_id.clone()
     }
 }
 
@@ -169,44 +190,16 @@ impl MessageFields for Notification {
     }
 }
 
-#[derive(Deserialize, Serialize, Debug)]
-pub struct Reconnect {
-    metadata: GenericMetadata,
-    payload: ReconnectSessionPayload,
-}
-
-impl Reconnect {
-    pub fn reconnect_url(&self) -> &String {
-        &self.payload.session.reconnect_url
-    }
-}
-
 impl MessageFields for Reconnect {
     fn id(&self) -> String {
         self.metadata.message_id.clone()
     }
 }
 
-#[derive(Deserialize, Serialize, Debug)]
-pub struct Revocation {
-    metadata: SubscriptionMetadata,
-    payload: SubscriptionPayload,
-}
-
 impl MessageFields for Revocation {
     fn id(&self) -> String {
         self.metadata.message_id.clone()
     }
-}
-
-#[derive(Serialize, Deserialize, Debug)]
-#[serde(untagged)]
-pub enum TwitchMessage {
-    Notification(Notification),
-    Welcome(Welcome),
-    Reconnect(Reconnect),
-    Revocation(Revocation),
-    Keepalive(Keepalive),
 }
 
 impl MessageFields for TwitchMessage {
